requirements/dev.txt
-r base.txt
pytest
pytest-django
drf-spectacular
drf-spectacular-sidecar
djangorestframework-simplejwt
